fix(v10/cloudflare): Keep the vite transform free of the wrangler import
wrangler 4.x needs Node 20+, and v10 still runs unit tests on Node 18.
